# backend/docs/swagger.yaml
basePath: /
definitions:
  github_com_zcl0621_compx576-smart-dairy-system_dto_alert.ListItem:
    properties:
      cow_id:
        type: string
      cow_name:
        type: string
      created_at:
        type: string
      id:
        type: string
      message:
        type: string
      metric_key:
        $ref: '#/definitions/model.MetricType'
      report_id:
        type: string
      resolved_at:
        type: string
      severity:
        $ref: '#/definitions/model.AlertSeverity'
      status:
        $ref: '#/definitions/model.AlertStatus'
      title:
        type: string
      updated_at:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_alert.ListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_alert.ListItem'
        type: array
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_alert.SummaryResponse:
    properties:
      active:
        type: integer
      critical:
        type: integer
      offline:
        type: integer
      warning:
        type: integer
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginResponse:
    properties:
      expires_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginUser'
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginUser:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.MessageResponse:
    properties:
      message:
        type: string
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetConfirmRequest:
    properties:
      new_password:
        type: string
      reset_token:
        type: string
    required:
    - new_password
    - reset_token
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetRequestEmailRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetVerifyRequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetVerifyResponse:
    properties:
      reset_token:
        type: string
    type: object
...
  model.AlertSeverity:
    enum:
    - warning
    - critical
    - offline
    type: string
    x-enum-varnames:
    - AlertSeverityWarning
    - AlertSeverityCritical
    - AlertSeverityOffline
  model.AlertStatus:
    enum:
    - active
    - resolved
    type: string
    x-enum-varnames:
    - AlertStatusActive
    - AlertStatusResolved
...
  model.MetricType:
    enum:
    - temperature
    - heart_rate
    - blood_oxygen
    - latitude
    - longitude
    - milk_amount
    - milking_duration
    - weight
    - device
    type: string
    x-enum-varnames:
    - MetricTypeTemperature
    - MetricTypeHeartRate
    - MetricTypeBloodOxygen
    - MetricTypeLatitude
    - MetricTypeLongitude
    - MetricTypeMilkAmount
    - MetricTypeMilkingDuration
    - MetricTypeWeight
    - MetricTypeDevice
...
  runtime_web_server_handler.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
info:
  contact: {}
  description: api for smart dairy cows
  title: smart dairy api
  version: "1.0"
paths:
  /api/alert/list:
    get:
      description: get paged alert list
      parameters:
      - default: 1
        description: page numb
        in: query
        name: page
        type: integer
      - default: 20
        description: page size
        in: query
        name: page_size
        type: integer
      - description: Cow ID
        in: query
        name: cow_id
        type: string
      - description: alert level
        in: query
        name: severity
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_alert.ListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: list alerts
      tags:
      - Alert
  /api/alert/summary:
    get:
      description: get active alert counts
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_alert.SummaryResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: get alert summary
      tags:
      - Alert
  /api/auth/login:
    post:
      consumes:
      - application/json
      description: login with email and password
      parameters:
      - description: login body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      summary: login
      tags:
      - Auth
  /api/auth/password-reset/confirm:
    post:
      consumes:
      - application/json
      description: reset password with one-time token
      parameters:
      - description: password reset confirm body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetConfirmRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      summary: confirm password reset
      tags:
      - Auth
  /api/auth/password-reset/request:
    post:
      consumes:
      - application/json
      description: send latest password reset code to email
      parameters:
      - description: password reset email body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetRequestEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      summary: request password reset code
      tags:
      - Auth
  /api/auth/password-reset/verify:
    post:
      consumes:
      - application/json
      description: verify reset code and issue one-time reset token
      parameters:
      - description: password reset verify body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.PasswordResetVerifyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      summary: verify password reset code
      tags:
      - Auth
  /api/auth/refresh:
    post:
      description: get new token for current user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_zcl0621_compx576-smart-dairy-system_dto_auth.LoginResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/runtime_web_server_handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: refresh token
      tags:
      - Auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"